Refactor: PEtab code to sub-package (cont.)
