+++ python/sdist/amici/parameter_mapping.py
@@ -1,4 +1,6 @@
 # some extra imports for backward-compatibility
+import warnings
+
 from .petab.conditions import (  # noqa # pylint: disable=unused-import
     fill_in_parameters,
     fill_in_parameters_for_condition,
@@ -14,4 +16,9 @@ from .petab.parameter_mapping import (  # noqa # pylint: disable=unused-import
     scale_parameters_dict,
     unscale_parameter,
     unscale_parameters_dict,
+)
+
+warnings.warn(
+    "Importing amici.parameter_mapping is deprecated. Use `amici.petab.parameter_mapping` instead.",
+    DeprecationWarning,
 )
